agri-connect-backend/requirements.txt
asgiref==3.9.1
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
gunicorn==23.0.0
joblib==1.5.1
mysqlclient==2.2.7
numpy==2.3.1
packaging==25.0
pandas==2.3.1
pillow==11.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.7.1
scipy==1.16.0
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
tzdata==2025.2
whitenoise==6.9.0
google-generativeai==0.7.2
